Trait bip_bencode::BListAccess

pub trait BListAccess<V> {
    fn get(&self, index: usize) -> Option<&V>;
    fn get_mut(&mut self, index: usize) -> Option<&mut V>;
    fn remove(&mut self, index: usize) -> Option<V>;
    fn insert(&mut self, index: usize, item: V);
    fn push(&mut self, item: V);
    fn len(&self) -> usize;
}

Trait for working with generic list data structures.

Required Methods

fn get(&self, index: usize) -> Option<&V>

Get a list element at the given index.

fn get_mut(&mut self, index: usize) -> Option<&mut V>

Get a mutable list element at the given index.

fn remove(&mut self, index: usize) -> Option<V>

Remove a list element at the given index.

fn insert(&mut self, index: usize, item: V)

Insert a list element at the given index.

fn push(&mut self, item: V)

Push an element to the back of the list.

fn len(&self) -> usize

Get the length of the list.